Move analysis_modudle.py to obsolete
